refactor(frontend): extract internal inspection alert
